feat: learnables

+++ 1-flashcard/backend/app/learnable/learnable.go
@@ -5,6 +5,6 @@ import "irayspace.com/flashcard/app/card"
 type Learnable struct {
 	ID           string    `json:"id"`
 	Card         card.Card `json:"card"`
-	LastReviewAt int       `json:"lastReviewAt"`
-	Interval     int       `json:"interval"`
+	LastReviewAt int64     `json:"lastReviewAt"`
+	Interval     int64     `json:"interval"`
 }
