+++ 1-flashcard/backend/app/session/session.go
@@ -5,5 +5,5 @@ import "irayspace.com/flashcard/app/learnable"
 type Session struct {
 	ID            string                `json:"id"`
 	Learnables    []learnable.Learnable `json:"learnables"`
-	LastSessionAt int                   `json:"lastSessionAt"`
+	LastSessionAt int64                 `json:"lastSessionAt"`
 }
